Fix and separate tests
